platform: refactor: move interrupt functions from platform.c
Moved platform_interrupt_* functions from platform.c to interrupt.c
to match CNL&APL. Updated headers for BYT&HSW to have the same functions
declarations as CNL&APL.
