+++ src/platform/haswell/include/platform/interrupt.h
@@ -66,14 +66,9 @@
 #define IRQ_MASK_EXT_PARITY	(1 << IRQ_NUM_EXT_PARITY)
 #define IRQ_MASK_TIMER3		(1 << IRQ_NUM_TIMER3)
 
-static inline void platform_interrupt_init(void) {}
+void platform_interrupt_init(void);
 
-/* haswell does not support child IRQs */
-static inline struct irq_desc *platform_irq_get_parent(uint32_t irq)
-{
-	return NULL;
-}
-
+struct irq_desc *platform_irq_get_parent(uint32_t irq);
 void platform_interrupt_set(int irq);
 void platform_interrupt_clear(uint32_t irq, uint32_t mask);
 uint32_t platform_interrupt_get_enabled(void);
