+++ src/platform/haswell/platform.c
@@ -159,72 +159,6 @@ int platform_boot_complete(uint32_t boot_message)
 	return 0;
 }
 
-void platform_interrupt_set(int irq)
-{
-	arch_interrupt_set(irq);
-}
-
-/* clear mask in PISR, bits are W1C in docs but some bits need preserved ?? */
-void platform_interrupt_clear(uint32_t irq, uint32_t mask)
-{
-	switch (irq) {
-	case IRQ_NUM_EXT_DMAC0:
-	case IRQ_NUM_EXT_DMAC1:
-	case IRQ_NUM_EXT_SSP0:
-	case IRQ_NUM_EXT_SSP1:
-		interrupt_clear(irq);
-		break;
-	default:
-		break;
-	}
-}
-
-/* TODO: expand this to 64 bit - should we just return mask of IRQ numbers */
-uint32_t platform_interrupt_get_enabled(void)
-{
-	return shim_read(SHIM_IMRD);
-}
-
-void platform_interrupt_mask(uint32_t irq, uint32_t mask)
-{
-	switch (irq) {
-	case IRQ_NUM_EXT_SSP0:
-		shim_write(SHIM_IMRD, SHIM_IMRD_SSP0);
-		break;
-	case IRQ_NUM_EXT_SSP1:
-		shim_write(SHIM_IMRD, SHIM_IMRD_SSP1);
-		break;
-	case IRQ_NUM_EXT_DMAC0:
-		shim_write(SHIM_IMRD, SHIM_IMRD_DMAC0);
-		break;
-	case IRQ_NUM_EXT_DMAC1:
-		shim_write(SHIM_IMRD, SHIM_IMRD_DMAC1);
-		break;
-	default:
-		break;
-	}
-}
-
-void platform_interrupt_unmask(uint32_t irq, uint32_t mask)
-{
-	switch (irq) {
-	case IRQ_NUM_EXT_SSP0:
-		shim_write(SHIM_IMRD, shim_read(SHIM_IMRD) & ~SHIM_IMRD_SSP0);
-		break;
-	case IRQ_NUM_EXT_SSP1:
-		shim_write(SHIM_IMRD, shim_read(SHIM_IMRD) & ~SHIM_IMRD_SSP1);
-		break;
-	case IRQ_NUM_EXT_DMAC0:
-		shim_write(SHIM_IMRD, shim_read(SHIM_IMRD) & ~SHIM_IMRD_DMAC0);
-		break;
-	case IRQ_NUM_EXT_DMAC1:
-		shim_write(SHIM_IMRD, shim_read(SHIM_IMRD) & ~SHIM_IMRD_DMAC1);
-		break;
-	default:
-		break;
-	}
-}
-
 /* init shim registers */
 static void platform_init_shim(void)
 {
